chore: move metrics to separate package
