+++ cmd/argo-watcher/prometheus/metrics.go
@@ -1,4 +1,4 @@
-package main
+package prometheus
 
 import (
 	"github.com/prometheus/client_golang/prometheus"
