+++ cmd/argo-watcher/prometheus/metrics_test.go
@@ -1,4 +1,4 @@
-package main
+package prometheus
 
 import (
 	"testing"
